requirements.txt
# Telegram Bot Framework
python-telegram-bot==20.7

# Database
aiosqlite==0.19.0

# Environment Variables
python-dotenv==1.0.0

# HTTP Clients
requests==2.31.0
httpx==0.25.2
urllib3==2.0.7

# Payment Gateway
yookassa==3.0.0
PyJWT>=2.8.0

# Retry Logic
tenacity==8.2.3

# RemnaWave panel API
remnawave>=2.0.0

# Webhook Server (ASGI)
quart>=0.19
hypercorn>=0.14

# Testing
pytest>=7.0
pytest-asyncio>=0.21
